Move CommutationChecker and Equivalence library to transpiler too
This commit moves the commutation checker and the equivalence library to
the new transpiler crate too. These structures while potentially more
general are in practice only ever used by the transpiler so in the
logical grouping of functionality it makes the most sense as part of the
transpiler crate. If we decide this doesn't make sense in the future we
can move either to qiskit-circuit or a different crate in the future.
But for right now this feels like the best match and also helps remove
the dependency of qiskit-transpiler on qiskit-accelerate.
