+++ crates/transpiler/src/equivalence.rs
