+++ crates/transpiler/src/lib.rs
@@ -10,6 +10,8 @@
 // copyright notice, and modified files need to carry a notice indicating
 // that they have been altered from the originals.
 
+pub mod commutation_checker;
+pub mod equivalence;
 pub mod passes;
 pub mod target;
 
@@ -19,3 +21,8 @@ pub mod target;
 // from the dependencies list.
 pub mod circuit_duration;
 pub mod twirling;
+
+use pyo3::import_exception_bound;
+
+import_exception_bound! {qiskit.exceptions, QiskitError}
+import_exception_bound! {qiskit.transpiler.exceptions, TranspilerError}
