+++ crates/transpiler/src/passes/commutation_analysis.rs
@@ -15,9 +15,9 @@ use pyo3::prelude::PyModule;
 use pyo3::{pyfunction, wrap_pyfunction, Bound, PyResult, Python};
 use qiskit_circuit::Qubit;
 
+use crate::commutation_checker::CommutationChecker;
 use hashbrown::HashMap;
 use pyo3::prelude::*;
-use qiskit_accelerate::commutation_checker::CommutationChecker;
 
 use pyo3::types::{PyDict, PyList};
 use qiskit_circuit::dag_circuit::{DAGCircuit, NodeType, Wire};
